Work around mixed-precision dtype mismatch in norm scaling

+++ lib/haliax/src/haliax/nn/normalization.py
@@ -114,9 +114,9 @@ class LayerNorm(LayerNormBase):
         out = out.astype(dtype)
 
         if self.weight is not None:
-            out = self.weight * out
+            out = self.weight.astype(out.dtype) * out
         if self.bias is not None:
-            out = out + self.bias
+            out = out + self.bias.astype(out.dtype)
         return out
 
 
@@ -134,9 +134,9 @@ class RmsNorm(LayerNormBase):
         out = out.astype(in_dtype)
 
         if self.weight is not None:
-            out = self.weight * out
+            out = self.weight.astype(out.dtype) * out
         if self.bias is not None:
-            out = out + self.bias
+            out = out + self.bias.astype(out.dtype)
         return out
